restart: systemd support; last channel as CLI arg.

+++ src/plugins/restartnotify/index.js
@@ -1,11 +1,17 @@
 const Plugin = require('../../structs/Plugin.js');
 const DatabasePlugin = require('../db');
 
+const lastRestartChannelCmd = process.argv
+    .map(arg => arg.split('='))
+    .find(arg => arg[0] === '--last-restart-channel');
+
 class RestartNotifyPlugin extends Plugin {
     static get deps() {
-        return [
-            DatabasePlugin
-        ];
+        return lastRestartChannelCmd ?
+            [] :
+            [
+                DatabasePlugin
+            ];
     }
 
     load() {
@@ -20,10 +26,14 @@ class RestartNotify {
     }
 
     async onReady() {
-        const channelId = await this.bot.db.get('lastRestartChannel');
-        if (!channelId) return;
-
-        this.bot.db.delete('lastRestartChannel');
+        let channelId;
+        if (lastRestartChannelCmd) {
+            channelId = lastRestartChannelCmd[1];
+        } else {
+            channelId = await this.bot.db.get('lastRestartChannel');
+            if (!channelId) return;
+            this.bot.db.delete('lastRestartChannel');
+        }
 
         const channel = await this.bot.client.channels.fetch(channelId);
         if (!channel) return;
